Test: Don't run API tests in frontend tests
There is not any value in doing this. It makes these tests slower, more
complex and opens it up to being flaky or broken. Let's keep it simple
and only do things, that are bringing real value.

Only commenting and not deleting so if we decide that we should be
exercising these we can easily add it back easily.